requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels-redis==4.2.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.4
django-redis==5.4.0
dlib==19.24.4
face-recognition==1.3.0
face_recognition_models==0.3.0
hyperlink==21.0.0
idna==3.6
imutils==0.5.4
incremental==22.10.0
msgpack==1.0.8
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
redis==5.0.3
scipy==1.13.0
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
zope.interface==6.2
